fix: refactor PdfPreview component
