+++ src/components/ModalUI.svelte
@@ -4,14 +4,10 @@
   import type { TConfig, ExportConfigModal } from "../modal";
   import { TFile } from "obsidian";
   import { settingToggle, settingDropdown, settingSlider, settingButton, settingDoubleText, icon } from "../actions";
-  import { getAllStyles, getPatchStyle, renderMarkdown, type ParamType } from "../render";
   import { exportToPDF, getOutputFile, getOutputPath } from "../pdf";
-  import { PageSize } from "../constant";
   import * as electron from "electron";
-  import { mm2px, safeParseFloat, px2mm, isNumber, safeParseInt } from "../utils";
-  import { fixDoc } from "../render";
-  import pLimit from "p-limit";
-  const fs = require("fs").promises;
+  import { isNumber } from "../utils";
+  import PdfPreview from "./PdfPreview.svelte";
 
   let {
     modal,
@@ -27,6 +23,9 @@
   let webviews = $state<electron.WebviewTag[]>([]);
   let lastPreview = $state<electron.WebviewTag | null>(null);
   let completed = $state(false);
+  let docs = $state<any[]>([]);
+
+  let pdfPreview: PdfPreview;
 
   const i18n = $derived(plugin.i18n);
   const settings = $derived(plugin.settings);
@@ -35,105 +34,12 @@
   let showCustomSize = $state(config.pageSize === "Custom");
   let showCustomMargin = $state(config.marginType === "3");
 
-  // Progress
-  let renderStates = $state<{ filename: string; status: number }[]>([]);
-  let docs = $state<any[]>([]);
-  let scale = $state(0.75);
-
-  export function calcPageSize(element?: HTMLDivElement) {
-    const { pageSize, pageWidth } = config;
-    const el = element ?? previewEl;
-    if (!el) return;
-    const width = PageSize?.[pageSize as string]?.[0] ?? safeParseFloat(pageWidth as string, 210);
-    scale = Math.floor((mm2px(width) / el.offsetWidth) * 100) / 100;
-  }
-
-  export function initRenderStates(data: ParamType[]) {
-    renderStates = data.map((param) => ({ status: 0, filename: param.file.name }));
-  }
-  export function updateRenderStates(i: number) {
-    renderStates[i].status = 1;
-  }
-
-  export async function calcWebviewSize() {
-    // @ts-ignore
-    await sleep(500);
-
-    webviews.forEach(async (e, i) => {
-      const [width, height] = await e.executeJavaScript("[document.body.offsetWidth, document.body.offsetHeight]");
-      docs[i].printSize = `${width}×${height}px²\n${px2mm(width)}×${px2mm(height)}mm²`;
-    });
-  }
-
-  async function renderFiles(data: ParamType[], docs?: any[], cb?: (i: number) => void) {
-    const concurrency = safeParseInt(settings.concurrency) || 5;
-    const limit = pLimit(concurrency);
-
-    const inputs = data.map((param, i) =>
-      limit(async () => {
-        const res = await renderMarkdown(param);
-        cb?.(i);
-        return res;
-      }),
-    );
-    let _docs = [...(docs ?? []), ...(await Promise.all(inputs))];
-
-    if (modal.file instanceof TFile) {
-      const leaf = modal.app.workspace.getLeaf();
-      await leaf.openFile(modal.file);
-    }
-
-    if (!modal.multiplePdf) {
-      _docs = modal.mergeDoc(_docs);
-    }
-    return _docs.map(({ doc, ...rest }) => {
-      return { ...rest, doc: fixDoc(doc, doc.title) };
-    });
-  }
-
-  export async function renderPreview(render = true) {
-    if (render) {
-      const { data, docs: allDocs } = await modal.getAllFiles();
-      initRenderStates(data);
-      docs = await renderFiles(data, allDocs, (i) => updateRenderStates(i));
-    }
-
-    webviews = [];
-    lastPreview = null;
-
-    const promises = docs.map((docItem) => {
-      return new Promise<void>((resolve) => {
-        // @ts-ignore
-        docItem.resolve = resolve;
-      });
-    });
-
-    await Promise.all(promises);
-    calcPageSize();
-    await calcWebviewSize();
-  }
-
-  function toggleTitle(value: boolean) {
-    webviews.forEach((wv, i) => {
-      wv.executeJavaScript(`
-        var _title = document.querySelector("h1.__title__");
-        if (_title) {
-          _title.style.display = "${value ? "block" : "none"}";
-        }
-      `);
-      const _title = docs[i]?.doc?.querySelector("h1.__title__") as HTMLHeadingElement;
-      if (_title) {
-        _title.style.display = value ? "block" : "none";
-      }
-    });
-  }
-
   export async function onCssSnippetChange() {
-    await renderPreview(false);
+    await pdfPreview?.renderPreview(false);
   }
 
   export async function refreshPreview() {
-    await renderPreview(true);
+    await pdfPreview?.renderPreview(true);
   }
 
   export async function handleExport() {
@@ -168,43 +74,6 @@
     }
   }
 
-  function initWebviewEvents(preview: any, docObj: any) {
-    webviews.push(preview);
-    lastPreview = preview; // keep track of the latest one
-
-    const handler = async () => {
-      completed = true;
-      getAllStyles().forEach(async (css) => {
-        await preview.insertCSS(css);
-      });
-      if (config.cssSnippet && config.cssSnippet != "0") {
-        try {
-          const cssSnippet = await fs.readFile(config.cssSnippet, { encoding: "utf8" });
-          const printCss = cssSnippet.replaceAll(/@media print\s*{([^}]+)}/g, "$1");
-          await preview.insertCSS(printCss);
-          await preview.insertCSS(cssSnippet);
-        } catch (error) {
-          console.warn(error);
-        }
-      }
-      await preview.executeJavaScript(modal.makeWebviewJs(docObj.doc));
-      getPatchStyle().forEach(async (css) => {
-        await preview.insertCSS(css);
-      });
-      if (docObj.resolve) {
-        docObj.resolve();
-      }
-    };
-
-    preview.addEventListener("dom-ready", handler);
-
-    return {
-      destroy() {
-        preview.removeEventListener("dom-ready", handler);
-      },
-    };
-  }
-
   // ── Page sizes ─────────────────────────────────────────
   const pageSizes = ["A0", "A1", "A2", "A3", "A4", "A5", "A6", "Legal", "Letter", "Tabloid", "Ledger", "Custom"];
   const pageSizeOptions = Object.fromEntries(pageSizes.map((s) => [s, s]));
@@ -221,23 +90,6 @@
   const hasSnippets = $derived(Object.keys(snippets).length > 0 && settings.enabledCss);
   const snippetOptions = $derived({ "0": "Not select", ...snippets });
 
-  // ── Preview area ref ───────────────────────────────────
-  let previewEl: HTMLDivElement;
-
-  onMount(() => {
-    const resizeObserver = new ResizeObserver(() => {
-      calcPageSize(previewEl);
-    });
-    resizeObserver.observe(previewEl);
-
-    // Initial render
-    renderPreview(true);
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  });
-
   function handleKeyup(event: KeyboardEvent) {
     if (event.key === "Enter") {
       handleExport();
@@ -247,42 +99,16 @@
 
 <div id="better-export-pdf">
   <!-- PDF Preview Area -->
-  <div class="pdf-preview">
-    <div class="progress">
-      {#if renderStates.length > 0 && !renderStates.every((item) => item.status)}
-        <div>Rendering...</div>
-        {#each renderStates as item}
-          <div>
-            {#if item.status}
-              <span use:icon={"check"}></span>
-            {:else}
-              <span use:icon={"loader"}></span>
-            {/if}
-            {item.filename}
-          </div>
-        {/each}
-      {/if}
-    </div>
-    <div bind:this={previewEl}>
-      {#each docs as item, i}
-        {#if modal.multiplePdf}
-          <div class="filename">{i + 1}-{item.doc.title}</div>
-        {/if}
-        <div class="webview-wrapper">
-          <div class="print-size" style:visibility={config.pageSize === "Custom" ? "visible" : "hidden"}>
-            {item.printSize ?? ""}
-          </div>
-          <webview
-            src="app://obsidian.md/help.html"
-            nodeintegration={true}
-            class="pdf-preview-webview"
-            style="--modal-scale: {scale};"
-            use:initWebviewEvents={item}
-          ></webview>
-        </div>
-      {/each}
-    </div>
-  </div>
+  <PdfPreview
+    {modal}
+    {plugin}
+    {config}
+    bind:webviews
+    bind:docs
+    bind:lastPreview
+    bind:completed
+    bind:this={pdfPreview}
+  />
 
   <!-- Settings Sidebar -->
   <!-- svelte-ignore a11y_no_static_element_interactions -->
@@ -295,7 +121,7 @@
         value: config.showTitle,
         onChange: (value) => {
           config.showTitle = value;
-          toggleTitle(value);
+          pdfPreview?.toggleTitle(value);
         },
       }}
     ></div>
@@ -309,8 +135,8 @@
         onChange: async (value) => {
           config.pageSize = value as TConfig["pageSize"];
           showCustomSize = value === "Custom";
-          calcPageSize();
-          await calcWebviewSize();
+          pdfPreview?.calcPageSize();
+          await pdfPreview?.calcWebviewSize();
         },
       }}
     ></div>
@@ -326,8 +152,8 @@
             isDebounce: true,
             onChange: async (value) => {
               config.pageWidth = value;
-              calcPageSize();
-              await calcWebviewSize();
+              pdfPreview?.calcPageSize();
+              await pdfPreview?.calcWebviewSize();
             },
           },
           input2: {
@@ -468,7 +294,7 @@
           value: config.cssSnippet ?? "0",
           onChange: async (value) => {
             config.cssSnippet = value;
-            await renderPreview(false);
+            await pdfPreview?.renderPreview(false);
           },
         }}
       ></div>
